requirements.txt
pytest
requests
allure-pytest
pydantic
playwright
pytest-playwright